+++ Microsoft.Toolkit.Services/Services/OneDrive/OneDriveService.cs
@@ -45,14 +45,14 @@ namespace Microsoft.Toolkit.Services.OneDrive
         public static OneDriveService Instance => _instance ?? (_instance = new OneDriveService());
 
         /// <summary>
-        /// Gets or sets AppClientId.
-        /// </summary>
-        protected string AppClientId { get; set; }
-
-        /// <summary>
         /// Gets or sets a value indicating whether service is initialized.
         /// </summary>
         protected static bool IsInitialized { get; set; }
+
+        /// <summary>
+        /// Gets or sets AppClientId.
+        /// </summary>
+        protected string AppClientId { get; set; }
 
         /// <summary>
         /// Gets or sets a value indicating whether user is connected.
@@ -121,6 +121,12 @@ namespace Microsoft.Toolkit.Services.OneDrive
         /// <returns>True or false.</returns>
         public virtual bool Initialize(string appClientId, string[] scopes, UIParent uiParent = null, string redirectUri = null)
         {
+#if WINRT
+            if (ServicePlatformInitializer == null)
+            {
+                ServicePlatformInitializer = new Uwp.OneDriveServicePlatformInitializer();
+            }
+#endif
             ServicePlatformService = ServicePlatformInitializer.CreateOneDriveServicePlatformInstance(this);
 
             AppClientId = appClientId;
