+++ Microsoft.Toolkit.Services/Services/OneDrive/Platform/Uwp/OneDriveServicePlatformInitializer.cs
@@ -10,10 +10,9 @@
 // THE CODE OR THE USE OR OTHER DEALINGS IN THE CODE.
 // ******************************************************************
 
-using Microsoft.Toolkit.Services.OneDrive;
 using Microsoft.Toolkit.Services.OneDrive.Platform;
 
-namespace Microsoft.Toolkit.Uwp.Services.OneDrive.Uwp
+namespace Microsoft.Toolkit.Services.OneDrive.Uwp
 {
     /// <summary>
     /// Platform implementation of initializer.
