+++ Microsoft.Toolkit.Uwp.SampleApp/SamplePages/OneDrive Service/FoldersPickerControl.xaml.cs
@@ -10,7 +10,6 @@
 // THE CODE OR THE USE OR OTHER DEALINGS IN THE CODE.
 // ******************************************************************
 
-using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Microsoft.Graph;
